+++ sort_helper.c
@@ -1,9 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   sort_helper.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: iarikupu <iarikupu@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/06/02 15:29:23 by iarikupu          #+#    #+#             */
+/*   Updated: 2022/06/02 15:40:00 by iarikupu         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "push_swap.h"
 
-int		find_min(t_stuff *all)
+int	find_min(t_stuff *all)
 {
-	int		min;
-	int     i;
+	int	min;
+	int	i;
 
 	i = 0;
 	min = all->a[0];
@@ -16,10 +28,10 @@ int		find_min(t_stuff *all)
 	return (min);
 }
 
-int		find_max(t_stuff *all)
+int	find_max(t_stuff *all)
 {
-	int		max;
-	int     i;
+	int	max;
+	int	i;
 
 	i = 0;
 	max = all->a[0];
@@ -32,9 +44,9 @@ int		find_max(t_stuff *all)
 	return (max);
 }
 
-int     get_pos(t_stuff *all, int n)
+int	get_pos(t_stuff *all, int n)
 {
-	int i;
+	int	i;
 
 	i = 0;
 	while (i < all->size_a)
@@ -48,7 +60,7 @@ int     get_pos(t_stuff *all, int n)
 
 void	best_move(t_stuff *all, int n, int i)
 {
-	if(i)
+	if (i)
 	{
 		if (all->size_a / 2 < get_pos(all, n))
 			ft_ra(all, 1);
@@ -58,7 +70,7 @@ void	best_move(t_stuff *all, int n, int i)
 	else
 	{
 		if (all->size_b / 2 < get_pos(all, n))
-			ft_rb(all,1);
+			ft_rb(all, 1);
 		else
 			ft_rrb(all, 1);
 	}
@@ -66,7 +78,7 @@ void	best_move(t_stuff *all, int n, int i)
 
 void	temp_sort(t_stuff *all, t_sort *s, int n)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	s->s = (int *)malloc((all->size_a) * sizeof(int));
